+++ arm7/source/common.h
@@ -21,3 +21,6 @@ extern ILogger* gLogger;
 #endif
 
 extern FATFS gFatFs;
+
+#define STRINGIFY2(x)  #x
+#define STRINGIFY(x)   STRINGIFY2(x)